remove check suites checks in github query
